add "/admin" to route .slip domain to login screen
Very small modification but it works. With current template, the user will be brought to a blank page. However when they are diverted properly to /admin path on the url it will fix this problem

ENV Variables 
Original App URL was SERVICE_FQDN_DIRECTUS_8055
I changed it to SERVICE_FQDN_DIRECTUS_8055=/admin

I can confirm it worked with my tests. For example, I used the template to create a slip domain. In the Directus Gui you will see the regular .slip domain without the /admin path, however, when you go to the generated domain, it will route you to the /admin path to the proper login screen.