fix(lang): split forgot_password to 2 translations
Until now, the translation of `auth.forgot_password_link` was used both as a link and as the header of the password recovery panel. In English, it still made sense, but in other languages, it simply sounds wrong. While adding the Polish language, I decided to separate these and also improve other translations, because the problem wasn’t limited to Polish.