fix(templates/supabase): Add missing SERVICE_URL_SUPABASEKONG environment variable definition
This variable is present when deploying Supabase on Coolify, but does not react to changes made to Kong's domain name. Currently, it will remain with the initially generated coolify domain name value and cannot be edited manually (regular env edit view grays the field out, dev mode ignores the changes).

Declaring it in the compose makes it react to the changes, just like `SERVICE_URL_SUPABASEKONG_8000` which was already setup accordingly.